Add a fallible UTF-8 decoder to util
This adds a UTF-8 decoder that takes an ArrayBufferView and throws an exception
on error. |util.decodeUtf8| is left as-is as it supports just passing in an
ArrayBuffer and will sanitize the output string rather than throwing an
exception. Profiling found that passing in a Uint8Array, then wrapping it in
another Uint8Array lead to an about 40% slowdown, so I've made
|util.decodeUtf8Array| explicitly take an ArrayBufferView which avoids this
pitfall.