Keep track of TextEditor focus in Alert transitions.
When transitioning from a Canvas Screen to an Alert Screen, keep track
of which TextEditor has focus. When returning to the Canvas Screen,
focus the TextEditor. This helps us match app expectations with regard
to keyboard visibility.

This patch also adds plumbing to allow RunTests tests to change the
current Displayable and compare the pixels of the current screen to a
reference image. These changes allow us to convert a bunch of gfx tests
to RunTests tests. The first converted test is AlertTest, which has been
updated to also test that focus is being tracked properly across showing
Alerts.