ensure MIDP.midletClassName defined before fs initialization
When we've compiled some classes ahead-of-time, main.js loads after
the compiled classes, which can be large and take a long time to load,
so fs initialization can race them.

But fs initialization depends on MIDP.midletClassName, since it adds
an extra directory (tcktestdir) if MIDP.midletClassName is set to RunTests.
And MIDP.midletClassName is currently set in main.js, which may not
be loaded by the time fs initialization begins.

So this change ensures that MIDP.midletClassName is defined before
fs initialization begins by setting it in midp.js instead of main.js.