Improve keyboard handling. Implement rotation.
Instead of creating our own thread on which to notify the listener of
keyboard visibility events, pass the VIRTUAL_KEYBOARD event to j2me and
process it in the `Display` class by notifying the listener there.

Improve heuristic for determining whether the keyboard is visible; we
now check against the expected inner and outer dimensions of the screen
(which we are able to determine on app launch) instead of against the last
known inner dimensions.

Handle device rotation by forwarding new canvas dimensions to j2me as a
native rotation event upon window resize. Note that the MIDlet will
never believe that it is rotated - only that it's dimensions have
changed. To have the MIDlet actually know about rotation, I think we
would have to apply a rotation to our canvas whenever the device
orientation changes.