Limit refresh rate to that of screen. Update tests
This prevents us from attempting to paint multiple times to the screen
when the screen can't actually keep up due to its refresh rate.

Update the asteroids test to use the accepted mechanism for doing 1
calculation per frame.

Add a benchmark test for `Graphics` performance. Current subtests
include `Graphics.drawRegion` and `Graphics.drawString`

Allocate the string "Async" once, outside of the functions it is used
in, and reference that string instead of using the string literal.