Optimize emoji handling in drawString().
This changeset rearranges drawString() so it allocates less: it removes
the `parts` array, and removes the `forEach` call, which avoids the
creation of a closure. In the case where the string contains no emoji,
it also avoids getting `g.currentFont` (which causes an allocation).

This reduces the cumulative JS allocations by almost 10% when running the
Asteroids midlet for a while.