Gfx fixes
Maintain information about what display is currently in the foreground,
and only allow drawing operations that are intended to go to the current
foreground display. This is part of the fix for #823.

Create a `ContextInfo` object that represents the translation, clipping,
font, and color information of a canvas context. Whenever a canvas
context gets modified, update its `ContextInfo`. Before any
drawing operation, check the `ContextInfo` and apply any values
from the `Graphics` object that don't match. This is the other part of
the fix for #823.

Fix the calculations in `renderRegion` and
`createImmutableImageDataRegion`. This affects situations where an icon
is drawn in various rotations (for example a single image that is
rotated frequently to create the effect of a progress throbber). Note
that `renderRegion` now draws to the temporary canvas so our memory
usage will be slightly higher. There might be a way around this if we
can figure out how to calculate the effect of the interaction of the
anchor type, rotation, and mirroring on the x and y coordinates of the
image we're drawing.