Remove unused `Graphics` methods
`Graphics.preserveMIDPRuntimeGC` and `Graphics.restoreMIDPRuntimeGC` are
package-private to the javax.microedition.lcdui package and are not
consumed by any members of that package. It seems we can safely remove
them from the `Graphics` class.

Note that this removes the declaration as well as the implementation;
any future errors are likely to be compilation failures rather than
runtime issues.