Fix objects returned from `Image.getGraphics`
Fixes mozilla/j2me.js#1383.

Fixes mozilla/j2me.js#1349.

For each canvas context (1 for every `ImageData` object and 1 for the
screen canvas), keep track of what `Graphics` object has been applied to
it. Whenever a property of the currently applied `Graphics` object
changes, remove the association of that object with the context since
they are no longer in sync.

Add a test that verifies that `Graphics` objects created from distinct
calls to `Image.getGraphics` on the same `Image` object do not interfere
with each other.

For each `Font`, keep a canvas context whose width and height are 0. Use
this context for text-measuring operations. This means that
`calcStringWidth` and related functions never have to change the `font`
property of a canvas context. `calcStringWidth` is called many times in
certain MIDlets, so the aggregate effect of this change should be
noticeable. Note that this performance is at the cost of some memory (1
canvas context per `Font` object).

Modify `withTextAnchor` to only change `textBaseline` once per
invocation, to never change `textAlign`, to return only an x-coordinate
(since it never actually does
anything with the y-coordinate), and to accept fewer arguments (only
those that are used).