update Isolate test expected output now that we don't set isolate ID redundantly
JVM.startIsolate0 used to set the isolate ID for the isolate it creates,
after which Isolate.registerNewIsolate would set it again (redundantly),
incrementing the internal counter within util.id() in the process.
Now, JVM.startIsolate0 doesn't set the ID, so the ID of that isolate (and others
created elsewhere) is one less than it used to be.

This commit updates the expected output of one of the Isolate tests to reflect
this change in isolate IDs.  In the process, it removes an obsolete file with an
old version of startIsolate0.