Implement an AMS MIDlet
Up to now, we've been telling the `MIDletSuiteLoader` class to launch
whatever user `MIDlet` we're trying to launch. We've been targeting the
background MIDlet of each suite we're running and then expecting the BG
MIDlet to call `BGUtils.setBGMIDletResident`. At that point, we've been
launching our best guess at what the FG MIDlet is (we currently assume
that there are only ever a max of 2 MIDlets in a suite).

This has a few important implications.

Because we launch the FG MIDlet when the BG MIDlet calls
`BGUtils.setBGMIDletResident`, and because that is frequently during the
BG MIDlet's `startApp` or constructor, the BG MIDlet does not get a
chance to fully initialize before the FG MIDlet starts up and takes the
foreground. This means that certain registrations that should be present
in the system by the time the FG MIDlet launches will not be present if
they were supposed to have been made by the BG MIDlet. It also means
that we cannot pause the BG MIDlet because the system will believe that
it is in a pre-startup state and will terminate it instead.

The MIDlet launched by `MIDletSuiteLoader` is supposed to be a special
"AMS" MIDlet that is responsible for managing user MIDlets
(launching/pausing/destroying, etc). The Isolate of the MIDlet launched
by `MIDletSuiteLoader` gets special privileges and is treated
specially by the MIDP system.

This patch implements an AMS MIDlet to be loaded into the special AMS
isolate. As a result of this patch we no longer load user MIDlets into
the AMS isolate. The AMS MIDlet implemented by this patch launches the
BG MIDlet if there is one specified in the JAD we are loading, then it
launches the FG MIDlet specified in the j2me.js configuration, or it
guesses like we've currently been doing if the configuration targets the
BG MIDlet.

The AMS MIDlet is now the handler for `NATIVE_MIDLET_EXECUTE_REQUEST`
events which we send to request MIDlet launches, and for
`NATIVE_MIDLET_DESTROY_REQUEST` which we now send to request MIDlet
destruction. These are implemented in terms of the `AmsUtil` and
`MIDletProxyList` classes which make certain methods available
only to the AMS MIDlet.