Notify threads based on priority instead of FIFO
ContextQueue is a prioritized queue of `Context` objects. Contexts are
returned from the ContextQueue based on their priority and how long they
have been waiting.

This patch implements ContextQueue and uses it to maintain prioritized
queues of contexts in lock objects. This way, when notify is called, we
give the lock to a context that makes sense due to its priority and how
long it has been waiting.