Fix race conditions in TestLocalMsgProtocol.
TestLocalMsgProtocol has some logic that looks like this:

    synchronized (someLockObject) {
        someFlag = true;
        someLockObject.notifyAll();
    }

    someBlockingFunctionThatMakesFlagActuallyTrue();

That's clearly a race condition, as the condition flag is set before the
condition is actually true. If we're relying on timing here, we're
giving ourselves a false sense of security.

Additionally, TestLocalMsgProtocol uses wait-loops to wait for
connections to be opened. However, it also opens those connections on
separate threads, so the wait-loops are unnecessary because we can just
`join` on the thread that creates the connection.